Create Code_DS3231_DHT22_SN75518
This code allow to use a DS3231 as a source of Real time Clock and also read a temperature from a DHT, now you can adapt it to show with your sequence. 
This code is related with a video on YouTube to allow explain the use of a driver VFD as interface a 7segments display.